requirements.txt
pytorch-lightning==1.8.6
tqdm
prettytable
menpo
mxnet
opencv-python
